refactor(llm): #1281 slice S1 — codec Protocol + CliNatsCodec rename + 2 new methods
- Promote LlmCodec to runtime_checkable Protocol in src/lyra/llm/codec.py (5 members).
- Rename concrete LlmCodec → CliNatsCodec via git mv (llm_codec.py → cli_nats_codec.py).
- Add encode_control(cmd) + set_session_store(store); move _CliSessionStore Protocol
  into cli_nats_codec.py (kept private — NATS-specific session bookkeeping).
- Update llm_client.py TYPE_CHECKING import to point at the new Protocol module.
- Update factory/llm_overlay.py to construct CliNatsCodec() (concrete) explicitly.
- Keep src/lyra/llm/llm_codec.py as backward-compat shim re-exporting CliNatsCodec
  as LlmCodec for two stragglers in tests/llm/ that tester-A migrates in Wave 3
  (T28). Shim removal is a follow-up after T28.

RED-GATE-S1: pyright clean across src/lyra/{llm,transport}/ +
isinstance(CliNatsCodec(), LlmCodec) holds (runtime_checkable Protocol).
